+++ src/App.js
@@ -1,101 +1,21 @@
 import React, { useEffect, useState } from "react";
-import { useQuery } from 'react-query'
-import axios from 'axios'
+import Home from "./pages/home";
+import { Routes, Route, Link } from 'react-router-dom'
+import About from "./pages/about";
 
 const App = () => {
-
-  const [urlBase, setUrlBase] = useState('https://swapi.dev/api/people/?page=1')
-  const [indexPage, setIndexPage] = useState(1)
-  const [arrayData, setArrayData] = useState();
-  // const [indexClick, setIndexClick] = useState(-1);
-  // const [isDisplay, setIsDisplay] = useState(false)
-  const [listIndexObjectOpen, setListIndexObjectOpen] = useState([]);
-  const { isLoading, error, data, isFetched } = useQuery(`page${indexPage}`, () =>
-    axios.get(urlBase)
-      .then(res => setArrayData(res.data))
-  )
-
-  const removeItemList = (index) => {
-    const newListItemRemoved = listIndexObjectOpen.filter(item => item != index)
-    // return newListItemRemoved
-    setListIndexObjectOpen(newListItemRemoved)
-  }
-
-  const removeAll = () => {
-    setListIndexObjectOpen([])
-  }
-
-  const addItemList = (index) => {
-    setListIndexObjectOpen([...listIndexObjectOpen, index])
-  }
-
-  useEffect(() => {
-    console.log(arrayData)
-  }, [isFetched])
-
-  const handleObjectDisplayOpen = (index) => {
-    // setIndexClick(index);
-    const isInclude = listIndexObjectOpen.includes(index)
-    isInclude ? removeItemList(index) : addItemList(index)
-  }
-
-  const nextUrlBase = () => {
-    setIndexPage(state => state + 1);
-    setUrlBase(arrayData.next)
-    removeAll()
-  }
-
-  const backUrlBase = () => {
-    setIndexPage(state => state - 1);
-    setUrlBase(arrayData.previous)
-    removeAll()
-  }
-
-  if (isLoading) return 'Carregando...';
-  if (error) return 'errro'
-
   return (
-    <div style={{ width: '100%', height: '100%', background: 'red' }}>
-      {isFetched && (
-        <div>
-          <h2>clique no nome para abrir as descrições</h2>
-          {arrayData.results.map((i, index) => {
-            return (
-              <div>
-                <p onClick={() => handleObjectDisplayOpen(index)} style={{cursor:'pointer'}}>{i.name}</p>
-                {
-                  listIndexObjectOpen.length > 0 && (
-                    listIndexObjectOpen.map(item => {
-                      return (
-                        item == index && (
-                          <div style={{ background: 'green' }}>
-                            <p>{i.gender}</p>
-                            <p>{i.birth_year}</p>
-                            <p>{i.height / 100}</p>
-                          </div>
-                        )
-                      )
-                    })
-                  )
-                }
-              </div>
-            )
-          })
-          }
-        </div>
-      )}
-
-      <input type='text' />
-      <div>
-        {
-          indexPage > 1 && (
-            <button onClick={backUrlBase} >Back</button>
-          )
-        }
-        <button onClick={nextUrlBase}>Next</button>
-      </div>
+    <div>
+      {/* <Link to="/">Home</Link>
+      <Link to="/about">About</Link> */}
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="about" element={<About />} >
+          <Route path=":person" element={<About />}/>
+        </Route>
+      </Routes>
     </div>
-  );
-}
 
+  )
+}
 export default App;
